+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web-ui-sdk
 
+## 1.4.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+
+## 1.4.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.44
+
 ## 1.4.41
 
 ### Patch Changes